:truck: rename: 조직 링크 도메인 분리에 따른 코드 위치 이동
