+++ src/main/java/com/sponus/sponusbe/domain/organization/entity/Organization.java
@@ -7,6 +7,7 @@ import com.sponus.sponusbe.domain.organization.dto.OrganizationUpdateRequest;
 import com.sponus.sponusbe.domain.organization.entity.enums.OrganizationStatus;
 import com.sponus.sponusbe.domain.organization.entity.enums.OrganizationType;
 import com.sponus.sponusbe.domain.organization.entity.enums.SuborganizationType;
+import com.sponus.sponusbe.domain.organizationLink.entity.OrganizationLink;
 import com.sponus.sponusbe.domain.tag.entity.Tag;
 import com.sponus.sponusbe.global.common.BaseEntity;
